Rename (to) TrayIconMenu
