+++ RP_Notify/Program.cs
@@ -8,7 +8,7 @@ using RP_Notify.PlayerWatchers.Foobar2000.BeefWebApiClient;
 using RP_Notify.RpApi;
 using RP_Notify.SongInfoListener;
 using RP_Notify.ToastHandler;
-using RP_Notify.TrayIcon;
+using RP_Notify.TrayIconMenu;
 using System;
 using System.Windows.Forms;
 
@@ -30,7 +30,7 @@ namespace RP_Notify
                 .AddSingleton<MusicBeeIPC>()
                 .AddSingleton<MusicBeeWatcher>()
                 .AddSingleton<ISongInfoListener, SongInfoListener.SongInfoListener>()
-                .AddSingleton<RpTrayIcon>()
+                .AddSingleton<RpTrayIconMenu>()
                 .AddSingleton<RpApplicationCore>();
 
             var serviceProvider = serviceCollection.BuildServiceProvider();
